+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 1.0.2
+- Updated image cdns.
 # 1.0.1
 - Added StarRailCharacterBuild#imageUrl.
 - Removed SkillTreeNode#previousNodeIds, please use SkillTreeNode#previousNodeId instead.